Refactor: 도메인 디렉토리 구조 변경
