+++ src/main/java/com/app/interstory/novel/domain/entity/CommentLike.java
@@ -1,4 +1,4 @@
-package com.app.interstory.novel.domain;
+package com.app.interstory.novel.domain.entity;
 
 import com.app.interstory.user.domain.entity.User;
 
@@ -9,22 +9,27 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "favorite_novel")
+@Table(name = "comment_like")
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 @Getter
-public class FavoriteNovel {
+@Builder
+public class CommentLike {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "favorite_novel_id")
-	private Long favoriteNovelId;
+	@Column(name = "comment_like_id")
+	private Long commentLikeId;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "user_id", nullable = false)
 	private User user;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "novel_id", nullable = false)
-	private Novel novel;
+	@JoinColumn(name = "comment_id", nullable = false)
+	private Comment comment;
+
+	public CommentLike(User user, Comment comment) {
+		this.user = user;
+		this.comment = comment;
+	}
 }
