+++ src/main/java/com/app/interstory/novel/domain/entity/Novel.java
@@ -1,5 +1,7 @@
-package com.app.interstory.novel.domain;
+package com.app.interstory.novel.domain.entity;
 
+import com.app.interstory.novel.domain.MainTag;
+import com.app.interstory.novel.domain.NovelStatus;
 import com.app.interstory.user.domain.entity.User;
 
 import jakarta.persistence.*;
