+++ src/main/java/com/app/interstory/novel/repository/CommentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.app.interstory.novel.domain.Comment;
+import com.app.interstory.novel.domain.entity.Comment;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
 	@Query("SELECT c FROM Comment c WHERE c.episode.episodeId = :episodeId ORDER BY "
