+++ src/main/java/com/app/interstory/novel/repository/EpisodeRepository.java
@@ -2,7 +2,7 @@ package com.app.interstory.novel.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.app.interstory.novel.domain.Episode;
+import com.app.interstory.novel.domain.entity.Episode;
 
 public interface EpisodeRepository extends JpaRepository<Episode, Long> {
 }
